Setup to include dist/ files as binary
The contents of the `dist/` can now be included in the repository as
binary files.  The `dist/` is still set to be ignored by git as it
should only be added when releasing a new version.  The `npm version`
command has been updated to force add `dist/` and commit it.